Hide label for user account menu (still shown to screen readers, hmm)
Ref #369

See core/themes/stable9/templates/block/block--system-menu-block.html.twig
for where Drupal makes the strange decision to show the label for the system
menu block to screenreaders even if the block label is set to none.

For other weirdness here i had to toggle this on and then off in the block settings
to get this to work.